Move the e2e tests to their own file
